Rename types to component
