+++ luda-adventure/src/app/game/types/game_object/component/renderer.rs
