+++ luda-adventure/src/app/game/types/game_object/mod.rs
@@ -1,9 +1,9 @@
+pub mod component;
 pub mod floor;
 pub mod player_character;
-pub mod types;
 pub mod wall;
 
+pub use component::*;
 pub use floor::*;
 pub use player_character::*;
-pub use types::*;
 pub use wall::*;
